+++ floor/writer.go
@@ -7,23 +7,26 @@ import (
 	"reflect"
 	"time"
 
+	"github.com/araddon/dateparse"
+	goparquet "github.com/fraugster/parquet-go"
 	"github.com/fraugster/parquet-go/floor/interfaces"
+	"github.com/fraugster/parquet-go/parquet"
 	"github.com/fraugster/parquet-go/parquetschema"
 	"github.com/pkg/errors"
-
-	goparquet "github.com/fraugster/parquet-go"
-	"github.com/fraugster/parquet-go/parquet"
 )
 
 // NewWriter creates a new high-level writer for parquet.
+// NOTE: We assume the schema definition is constant.
 func NewWriter(w *goparquet.FileWriter) *Writer {
 	return &Writer{
-		w: w,
+		w:         w,
+		schemaDef: w.GetSchemaDefinition(),
 	}
 }
 
 // NewFileWriter creates a nigh high-level writer for parquet
 // that writes to a particular file.
+// NOTE: We assume the schema definition is constant.
 func NewFileWriter(file string, opts ...goparquet.FileWriterOption) (*Writer, error) {
 	f, err := os.OpenFile(file, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0644)
 	if err != nil {
@@ -31,30 +34,30 @@ func NewFileWriter(file string, opts ...goparquet.FileWriterOption) (*Writer, er
 	}
 
 	w := goparquet.NewFileWriter(f, opts...)
-
 	return &Writer{
-		w: w,
-		f: f,
+		w:         w,
+		f:         f,
+		schemaDef: w.GetSchemaDefinition(),
 	}, nil
 }
 
 // Writer represents a high-level writer for parquet files.
 type Writer struct {
-	w *goparquet.FileWriter
-	f io.Closer
+	w         *goparquet.FileWriter
+	f         io.Closer
+	schemaDef *parquetschema.SchemaDefinition
 }
 
 // Write adds a new object to be written to the parquet file. If
 // obj implements the floor.Marshaller object, then obj.(Marshaller).Marshal
-// will be called to determine the data, otherwise reflection will be
-// used.
+// will be called to determine the data, otherwise reflection will be used.
 func (w *Writer) Write(obj interface{}) error {
 	m, ok := obj.(interfaces.Marshaller)
 	if !ok {
-		m = &reflectMarshaller{obj: obj, schemaDef: w.w.GetSchemaDefinition()}
+		m = &reflectMarshaller{obj: obj, schemaDef: w.schemaDef}
 	}
 
-	data := interfaces.NewMarshallObjectWithSchema(nil, w.w.GetSchemaDefinition())
+	data := interfaces.NewMarshallObjectWithSchema(nil, w.schemaDef)
 	if err := m.MarshalParquet(data); err != nil {
 		return err
 	}
@@ -76,8 +79,33 @@ func (m *reflectMarshaller) MarshalParquet(record interfaces.MarshalObject) erro
 }
 
 func (m *reflectMarshaller) marshal(record interfaces.MarshalObject, value reflect.Value, schemaDef *parquetschema.SchemaDefinition) error {
-	if err := m.decodeStruct(record, value, schemaDef); err != nil {
-		return err
+	if value.Type().Kind() == reflect.Ptr {
+		if value.IsNil() {
+			return errors.New("object is nil")
+		}
+		value = value.Elem()
+	}
+
+	typ := value.Type()
+
+	if typ.Kind() == reflect.Struct {
+		return m.decodeStruct(record, value, schemaDef)
+	}
+
+	if typ.Kind() != reflect.Map {
+		return fmt.Errorf("object needs to be a struct, *struct or map, it's a %v instead", typ)
+	}
+
+	iter := value.MapRange()
+	for iter.Next() {
+		fieldName := iter.Key().String()
+		subSchemaDef := schemaDef.SubSchema(fieldName)
+		field := record.AddField(fieldName)
+
+		err := m.decodeValue(field, iter.Value(), subSchemaDef)
+		if err != nil {
+			return err
+		}
 	}
 
 	return nil
@@ -164,7 +192,7 @@ func (m *reflectMarshaller) decodeValue(field interfaces.MarshalElement, value r
 		return nil
 	}
 
-	if value.Kind() == reflect.Ptr {
+	if value.Kind() == reflect.Ptr || value.Kind() == reflect.Interface {
 		if value.IsNil() {
 			return nil
 		}
@@ -193,6 +221,20 @@ func (m *reflectMarshaller) decodeValue(field interfaces.MarshalElement, value r
 		}
 	}
 
+	if !elem.IsSetType() && !elem.IsSetConvertedType() && elem.GetNumChildren() > 0 && value.Kind() == reflect.Map {
+		group := field.Group()
+		iter := value.MapRange()
+		for iter.Next() {
+			fieldName := iter.Key().String()
+			err := m.decodeValue(group.AddField(fieldName), iter.Value(), schemaDef.SubSchema(fieldName))
+			if err != nil {
+				return err
+			}
+		}
+
+		return nil
+	}
+
 	switch elem.GetType() {
 	case parquet.Type_INT64:
 		switch value.Kind() {
@@ -214,6 +256,45 @@ func (m *reflectMarshaller) decodeValue(field interfaces.MarshalElement, value r
 			return errors.Errorf("unable to decode %s:%s to int32", elem.Name, value.Kind())
 		}
 		return nil
+	case parquet.Type_INT96:
+		switch value.Kind() {
+		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
+			return m.decodeUnixTime(field, value.Int())
+		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
+			return m.decodeUnixTime(field, int64(value.Uint()))
+		case reflect.String:
+			dt, _ := dateparse.ParseAny(value.String())
+			field.SetInt96(goparquet.TimeToInt96(dt))
+			return nil
+		case reflect.Slice:
+			if value.IsNil() {
+				return nil
+			}
+			if value.Type().Elem().Kind() != reflect.Uint8 {
+				return fmt.Errorf("field is of type INT96 but type is %s", value.Type().String())
+			}
+
+			if value.Len() != 12 {
+				return fmt.Errorf("field is of type INT96 but length is %d", value.Len())
+			}
+			var dst [12]byte
+			src := value.Interface().([]byte)
+			copy(dst[:], src)
+			field.SetInt96(dst)
+			return nil
+		case reflect.Array:
+			if value.Type().Elem().Kind() != reflect.Uint8 {
+				return fmt.Errorf("field is of type INT96 but type is %s", value.Type().String())
+			}
+			if value.Len() != 12 {
+				return fmt.Errorf("field is of type INT96 but length is %d", value.Len())
+			}
+			var dst [12]byte
+			src := value.Interface().([12]byte)
+			copy(dst[:], src[:])
+			field.SetInt96(dst)
+			return nil
+		}
 	}
 
 	switch value.Kind() {
@@ -243,6 +324,31 @@ func (m *reflectMarshaller) decodeValue(field interfaces.MarshalElement, value r
 	}
 }
 
+func (m *reflectMarshaller) decodeUnixTime(field interfaces.MarshalElement, i64 int64) error {
+	// best effort parse unix timestamps.
+	// since 99% of the time these are timestamps and are <= now this is a fairly safe bet
+	digits := i64Digits(i64)
+	now := time.Now()
+
+	switch {
+	case digits <= i64Digits(now.Unix()):
+		dt := time.Unix(i64, 0)
+		field.SetInt96(goparquet.TimeToInt96(dt))
+	case digits <= i64Digits(now.UnixNano()/1000000):
+		dt := time.Unix(0, i64*int64(time.Millisecond))
+		field.SetInt96(goparquet.TimeToInt96(dt))
+	case digits <= i64Digits(now.UnixNano()/1000):
+		dt := time.Unix(0, i64*int64(time.Microsecond))
+		field.SetInt96(goparquet.TimeToInt96(dt))
+	case digits <= i64Digits(now.UnixNano()):
+		dt := time.Unix(0, i64)
+		field.SetInt96(goparquet.TimeToInt96(dt))
+	default:
+		return fmt.Errorf("field is of type INT96 but value is not valid %d", i64)
+	}
+	return nil
+}
+
 func (m *reflectMarshaller) decodeByteSliceOrArray(field interfaces.MarshalElement, value reflect.Value, schemaDef *parquetschema.SchemaDefinition) error {
 	elem := schemaDef.SchemaElement()
 	if elem == nil {
@@ -266,20 +372,8 @@ func (m *reflectMarshaller) decodeByteSliceOrArray(field interfaces.MarshalEleme
 		}
 		field.SetByteArray(value.Bytes())
 	case reflect.Array:
-		if elem.GetType() == parquet.Type_INT96 {
-			if value.Len() != 12 {
-				return fmt.Errorf("field is of type INT96 but length is %d", value.Len())
-			}
-			data := reflect.New(value.Type()).Elem()
-			_ = reflect.Copy(data, value)
-
-			field.SetInt96(data.Interface().([12]byte))
-			return nil
-		}
 		data := reflect.MakeSlice(reflect.TypeOf([]byte{}), value.Len(), value.Len())
-
 		_ = reflect.Copy(data, value)
-
 		field.SetByteArray(data.Bytes())
 	}
 	return nil
@@ -358,4 +452,13 @@ func (w *Writer) Close() error {
 	}
 
 	return w.w.Close()
+}
+
+func i64Digits(number int64) int {
+	count := 0
+	for number != 0 {
+		number /= 10
+		count++
+	}
+	return count
 }