+++ CHANGELOG.md
@@ -7,7 +7,35 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Nothing yet
+## [v0.9.0] - 2022-02-10
+
+- Implemented lazy loading of pages.
+- Added support for writing multiple data pages per row group.
+
+## [v0.8.0] - 2022-02-02
+- Set correct total size and total compressed in row group data.
+- Fixed delta bit-pack encoding of int32 and int64 and delta-length bit-packing encoding of byte\_arrays when only a single value is written.
+- Fixed build issue in the floor package that somehow made it into the previous release.
+
+## [v0.7.0] - 2022-02-01
+- Relax schema parser requirement that field names begin with a letter or underscore.
+- Remove unsigned support from all integer encodings.
+- Introduced explicit Clone method for parquetschema.SchemaDefinition.
+- Corrected parsing of DECIMAL as ConvertedType if scale and precision aren't available.
+- Fixed nested optional field reads.
+
+## [v0.6.1] - 2021-11-19
+- perf - cache result of GetSchemaDefinition call
+- updated readme, added special mentions section
+
+## [v0.6.0] - 2021-11-3
+- Added a schema generator which uses reflection to automatically generate a parquet schema from a go object.
+- Upgraded CI golang to 1.17.2
+- Added test to reproduce issue 41: https://github.com/fraugster/parquet-go/issues/41 
+
+## [v0.5.0] - 2021-10-29
+- Upgraded thrift to v0.15.0
+- Fixed reading & writing of INT96 time
 
 ## [v0.4.0] - 2021-10-06
 - Fixed issues where fields in structs that were not defined in the schema raised errors (array, slice, time, map)
@@ -56,7 +84,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [v0.1.0] - 2020-04-24
 - Initial release
 
-[Unreleased]: https://github.com/fraugster/parquet-go/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/fraugster/parquet-go/compare/v0.9.0...HEAD
+[v0.9.0]: https://github.com/fraugster/parquet-go/releases/tag/v0.9.0
+[v0.8.0]: https://github.com/fraugster/parquet-go/releases/tag/v0.8.0
+[v0.7.0]: https://github.com/fraugster/parquet-go/releases/tag/v0.7.0
+[v0.6.1]: https://github.com/fraugster/parquet-go/releases/tag/v0.6.1
+[v0.6.0]: https://github.com/fraugster/parquet-go/releases/tag/v0.6.0
+[v0.5.0]: https://github.com/fraugster/parquet-go/releases/tag/v0.5.0
 [v0.4.0]: https://github.com/fraugster/parquet-go/releases/tag/v0.4.0
 [v0.3.0]: https://github.com/fraugster/parquet-go/releases/tag/v0.3.0
 [v0.2.1]: https://github.com/fraugster/parquet-go/releases/tag/v0.2.1